Move GeneExclusions into pipeline_utils and setup v2
